Temporarily (?) remove CNAME file
The http-kit.org domain has expired [1], causing
http-kit.github.com to stop working.

It may be nice to renew the domain, but it's currently still
being held by the registrar. Am removing the CNAME file here
so that folks can at least access the content via http-kit.github.com
so long.

If/when the .org domain is released by the registrar, we can
decide to either renew the domain and revert this commit - or
just let the domain go and use http-kit.github.com as the new
canonical project URL.

[1] https://github.com/http-kit/http-kit/issues/424